Add line limit to ex_unit with default 100
This option defaults to 100. This creates an easier to look output of tests that show in glance
an approximation of the number of tests run.

This also helps with consoles/editors that have problems with very displaying very long
lines. Though this might seem a bit odd, Emacs is one of those editors that has an issue with very
long lines.